feature: added namespace seal status endpoint
