+++ helper/namespace/namespace.go
@@ -29,6 +29,9 @@ var reservedNames = []string{
 	"auth",
 	"cubbyhole",
 	"identity",
+	"seal",
+	"unseal",
+	"seal-status",
 }
 
 type (
