+++ vault/seal_manager.go
@@ -182,9 +182,14 @@ func (sm *SealManager) SecretProgress(ns *namespace.Namespace, lock bool) (int, 
 }
 
 func (sm *SealManager) GetSealStatus(ctx context.Context, ns *namespace.Namespace, lock bool) (*SealStatusResponse, error) {
+	// Verify that any kind of seal exists for a namespace
+	seals, ok := sm.sealsByNamespace[ns.UUID]
+	if !ok {
+		return nil, errors.New("namespace is not set with a seal")
+	}
+
 	// Check the barrier first
 	barrier := sm.NamespaceBarrier(ns.Path)
-
 	init, err := barrier.Initialized(ctx)
 	if err != nil {
 		sm.logger.Error("namespace barrier init check failed", "namespace", ns.Path, "error", err)
@@ -195,8 +200,8 @@ func (sm *SealManager) GetSealStatus(ctx context.Context, ns *namespace.Namespac
 		return nil, nil
 	}
 
-	seal := *sm.sealsByNamespace[ns.UUID]["default"]
 	// Verify the seal configuration
+	seal := *seals["default"]
 	sealConf, err := seal.BarrierConfig(ctx, ns)
 	if err != nil {
 		return nil, err
